refactor: remove unused type definitions and add new Prisma schema files
