+++ prisma/schema/survey.prisma
@@ -1,0 +1,38 @@
+model CustomerSatisfactionSurvey {
+  id         String   @id @default(ulid())
+  orderId    String   @unique
+  categoryId String
+  submitDate DateTime @default(now())
+  question1  Int
+  question2  Int
+  question3  Int
+  question4  Int
+  comments   String?
+  createdAt  DateTime @default(now())
+  updatedAt  DateTime @updatedAt
+
+  // Relationships
+  order    Order           @relation(fields: [orderId], references: [id])
+  category SurveyCategory  @relation(fields: [categoryId], references: [id])
+
+  Order    Order[]         @relation("OrderToSurvey")
+
+  @@index([categoryId], name: "idx_survey_categoryId")
+}
+
+model SurveyCategory {
+  id          String    @id @default(ulid())
+  isDeleted   Boolean   @default(false)
+  name        String
+  description String?
+  question1   String
+  question2   String
+  question3   String
+  question4   String
+  surveys     CustomerSatisfactionSurvey[]
+  createdAt   DateTime  @default(now())
+  updatedAt   DateTime  @updatedAt
+
+  @@index([name], name: "idx_surveycategory_name")
+  @@index([isDeleted], name: "idx_survey-category_isDeleted")
+}
