--- src/types/Category.d.ts
@@ -1,7 +1,0 @@
-export type Category = {
-  id: string
-  name: string
-  description?: string | null
-  createdAt: Date
-  updatedAt: Date
-}
