--- src/types/CustomerSatisfactionSurveyCategory.d.ts
@@ -1,11 +1,0 @@
-export type CustomerSatisfactionSurveyCategory = {
-  id: string
-  name: string
-  description?: string | null
-  question1: string
-  question2: string
-  question3: string
-  question4: string
-  createdAt: Date
-  updatedAt: Date
-}
