--- src/types/Log.d.ts
@@ -1,9 +1,0 @@
-export type Log = {
-  id: string
-  userId?: string | null
-  createdDate: Date
-  createdById?: string | null
-  reason: string
-  systemText: string
-  userText: string
-}
