--- src/types/Misc.d.ts
@@ -1,4 +1,0 @@
-declare type LinkType = {
-  href: string
-  displayName: string
-}
